fix: address Laravel Doctor diagnostics
