+++ tests/Unit/Models/Concerns/IsEmployableTest.php
@@ -16,66 +16,10 @@ namespace Tests\Unit\Models\Concerns;
 use App\Models\Concerns\IsEmployable;
 use App\Models\Contracts\Employable;
 use App\Models\Lifecycle\Employment;
-use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\MorphMany;
 use Illuminate\Database\Eloquent\Relations\MorphOne;
-use Illuminate\Database\Query\Builder as QueryBuilder;
-use JMac\Testing\Double;
-
-/** @extends EloquentBuilder<Employment> */
-final class FakeEmploymentBuilder extends EloquentBuilder {}
-
-/** @implements Employable<self> */
-final class EmploymentStateModel extends Model implements Employable
-{
-    /** @use IsEmployable<self> */
-    use IsEmployable;
-
-    public bool $futureEmploymentExists = false;
-
-    public bool $currentEmploymentExists = false;
-
-    public bool $employmentExists = false;
-
-    /** @return MorphOne<Employment, self> */
-    public function futureEmployment(): MorphOne
-    {
-        return $this->employmentHasOne($this->futureEmploymentExists);
-    }
-
-    /** @return MorphOne<Employment, self> */
-    public function currentEmployment(): MorphOne
-    {
-        return $this->employmentHasOne($this->currentEmploymentExists);
-    }
-
-    /** @return MorphMany<Employment, self> */
-    public function employments(): MorphMany
-    {
-        $builder = $this->employmentBuilder($this->employmentExists);
-
-        return new MorphMany($builder, new self(), 'employable_type', 'employable_id', 'id');
-    }
-
-    /** @return MorphOne<Employment, self> */
-    private function employmentHasOne(bool $exists): MorphOne
-    {
-        $builder = $this->employmentBuilder($exists);
-
-        return new MorphOne($builder, new self(), 'employable_type', 'employable_id', 'id');
-    }
-
-    private function employmentBuilder(bool $exists): FakeEmploymentBuilder
-    {
-        $query = Double::for(QueryBuilder::class);
-        $query->expects('exists')->returns($exists);
-        $builder = new FakeEmploymentBuilder($query);
-        $builder->setModel(new Employment());
-
-        return $builder;
-    }
-}
+use Tests\Unit\Models\Concerns\Support\EmploymentStateModel;
 
 describe('IsEmployable Trait Unit Tests', function () {
     describe('employment relationships', function () {
