+++ tests/Unit/Models/Concerns/IsRetirableTest.php
@@ -7,56 +7,10 @@ namespace Tests\Unit\Models\Concerns;
 use App\Models\Concerns\IsRetirable;
 use App\Models\Contracts\Retirable;
 use App\Models\Lifecycle\Retirement;
-use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\MorphMany;
 use Illuminate\Database\Eloquent\Relations\MorphOne;
-use Illuminate\Database\Query\Builder as QueryBuilder;
-use JMac\Testing\Double;
-
-/** @extends EloquentBuilder<Retirement> */
-final class FakeRetirementBuilder extends EloquentBuilder {}
-
-/** @implements Retirable<self> */
-final class RetirementStateModel extends Model implements Retirable
-{
-    /** @use IsRetirable<self> */
-    use IsRetirable;
-
-    public bool $currentRetirementExists = false;
-
-    /** @return MorphOne<Retirement, self> */
-    public function currentRetirement(): MorphOne
-    {
-        return $this->retirementHasOne($this->currentRetirementExists);
-    }
-
-    /** @return MorphMany<Retirement, self> */
-    public function retirements(): MorphMany
-    {
-        $builder = $this->retirementBuilder(false);
-
-        return new MorphMany($builder, new self(), 'retirable_type', 'retirable_id', 'id');
-    }
-
-    /** @return MorphOne<Retirement, self> */
-    private function retirementHasOne(bool $exists): MorphOne
-    {
-        $builder = $this->retirementBuilder($exists);
-
-        return new MorphOne($builder, new self(), 'retirable_type', 'retirable_id', 'id');
-    }
-
-    private function retirementBuilder(bool $exists): FakeRetirementBuilder
-    {
-        $query = Double::for(QueryBuilder::class);
-        $query->expects('exists')->returns($exists);
-        $builder = new FakeRetirementBuilder($query);
-        $builder->setModel(new Retirement());
-
-        return $builder;
-    }
-}
+use Tests\Unit\Models\Concerns\Support\RetirementStateModel;
 
 describe('IsRetirable', function () {
     test('provides polymorphic retirement relationships', function () {
